add yaml to mock modules
